cpu: initial decoder implementation.
The way this work is that the decoder decodes the instruction into an
Instruction, which contains an opcode and operand enum. The CPU can
then take this Instruction and execute it according to the opcode and
operand enum. This helps to separate the execution and disassemble
stage. Also this allows the decoder to be a totally separate stage.