cpu: Initial Decoder implementation.
The Decoder acts as the instruction decoder for the CPU. The decoder has
access to the CPU's memory and instruction pointer.

To be cycle-accurate, the decoder does not decode a instruction right
away. Instead, on the first cycle, the decoder first fetches the opcode,
then decides if there's an immediate following the opcode. If that's the
case, then on the next cycle (or two cycles if the immediate is
16-bit), the decoder fetches the immediate, then
return the complete instruction. If not, then the instruction is
returned right away. The instruction is returned as a Instruction
object (surprise!).