cpu/cpu: hook it with MicroOpDecoder and initial micro op execution
Now, whenever the Decoder finishes decoding an instruction, that
instruction is passed to MicroOpDecoder, which further decodes it
into one or multiple micro ops. These micro ops are the basic
execution unit of the CPU. Decoded micro ops are pushed into a queue
to be executed, until the queue is empty, in which the CPU instructs
the Decoder to start decoding the next instruction, and the loop repeats.