memory/MBC1: fix unittest to check the last element in a ROM/RAM bank
The loop to check the element values in a ROM/RAM bank skips the
last element because the condition is `(i < ROM_SWITCHABLE_END)`,
while ROM_SWITCHABLE_END is the address of the last element in the
bank. Fix it by changing the condition to `<=`. Also fix the same
issue with RAM_EXTERNAL_END.