memory/MemoryAddrProxy: implement operator*()
In the case of assigning a MemoryAddrProxy to another MemoryAddrProxy,
for example `mem[0x1111] = mem[0x2222]`, the right operand must be casted
to a MemoryValue so its value can be assigned to the left operand.
That might look cumbersome, so now "dereferencing" the right operand
using the * operator also returns its value at that address. Inspired
by std::optional.